+++ scan2.sh
@@ -1,0 +1,3 @@
+rg -n "\bmalloc\(" src/
+rg -n "\bnew\b" src/
+rg -n "\bfopen\(" src/